Readd muzzle flash effect through data.
